restructured the gem a bit and added a few specs
