+++ lib/active_interactor/interactor/core.rb
@@ -2,31 +2,75 @@
 
 module ActiveInteractor
   module Interactor
-    # Provides ActiveInteractor::Interactor methods to included classes
-    #
-    # @author Aaron Allen <hello@aaronmallen.me>
-    # @since 0.0.1
-    # @version 0.1
-    #
-    # @!attribute [r] context
-    #  @return [ActiveInteractor::Context::Base] an instance of {ActiveInteractor::Context::Base}
-    module Core
-      extend ActiveSupport::Concern
+    module ClassMethods
+      # Invoke an interactor. This is the primary public API method to an
+      #  interactor.
+      #
+      # @example Run an interactor
+      #  MyInteractor.perform(name: 'Aaron')
+      #  #=> <#MyInteractor::Context name='Aaron'>
+      #
+      # @param context [Hash] properties to assign to the interactor context
+      # @return [ActiveInteractor::Context::Base] an instance of context
+      def perform(context = {})
+        new(context).tap(&:call_perform).context
+      end
 
-      included do
-        include Action
-        include Callbacks
-        include Context
-
-        attr_reader :context
-
-        # A new instance of {Core}
-        # @param context [Hash, nil] the properties of the context
-        # @return [Core] a new instance of {Core}
-        def initialize(context = {})
-          @context = self.class.context_class.new(self, context)
-        end
+      # Invoke an Interactor. The {.perform!} method behaves identically to
+      #  the {.perform} method with one notable exception. If the context is failed
+      #  during invocation of the interactor, the {ActiveInteractor::Context::Failure}
+      #  is raised.
+      #
+      # @example Run an interactor
+      #  MyInteractor.perform!(name: 'Aaron')
+      #  #=> <#MyInteractor::Context name='Aaron'>
+      #
+      # @param context [Hash] properties to assign to the interactor context
+      # @return [ActiveInteractor::Context::Base] an instance of context
+      def perform!(context = {})
+        new(context).tap(&:call_perform!).context
       end
     end
+
+    # Calls {#call_perform!} and rescues {ActiveInteractor::Context::Failure}
+    # @private
+    def call_perform
+      call_perform!
+    rescue ActiveInteractor::Context::Failure => exception
+      ActiveInteractor.logger.error("ActiveInteractor: #{exception}")
+    end
+
+    # Calls {#perform} with callbacks and context validation
+    # @private
+    def call_perform!
+      run_callbacks :perform do
+        context.called!(self)
+        fail_on_invalid_context!(:calling)
+        perform
+        fail_on_invalid_context!(:called)
+        context.clean! if should_clean_context?
+      rescue # rubocop:disable Style/RescueStandardError
+        context.rollback!
+        raise
+      end
+    end
+
+    # Calls {#rollback} with callbacks
+    # @private
+    def call_rollback
+      run_callbacks :rollback do
+        rollback
+      end
+    end
+
+    # Invoke an Interactor instance without any hooks, tracking, or rollback
+    # @abstract It is expected that the {#perform} method is overwritten
+    #  for each interactor class.
+    def perform; end
+
+    # Reverse prior invocation of an Interactor instance.
+    # @abstract Any interactor class that requires undoing upon downstream
+    #  failure is expected to overwrite the {#rollback} method.
+    def rollback; end
   end
 end
