--- spec/integration_spec.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-RSpec.describe 'Integration' do
-end
