+++ spec/spec_helper.rb
@@ -26,4 +26,4 @@ RSpec.configure do |config|
   Kernel.srand config.seed
 end
 
-Dir[File.expand_path('support/*.rb', __dir__)].each { |f| require f }
+Dir[File.join(__dir__, 'support', '**', '*.rb')].each { |f| require f }
